# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Booking API
  description: This is a sample API for booking management.
  version: 1.0.0
paths:
  /api/booking/search:
    post:
      summary: Search for bookings
      description: Retrieve a booking based on search criteria.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bookingCode:
                  type: string
                  pattern: "^[A-Z0-9]{6}$"
                familyName:
                  type: string
                  example: "HESP"
              additionalProperties: false
              required:
              - bookingCode
              - familyName

      responses:
        '200':
          description: A booking based on search criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResult'
        '400':
          description: Invalid request parameters
        '404':
          description: Booking not found based on parameters
        '500':
          description: Internal server error
  
  /api/booking/kill:
    post:
      summary: Set the healthy flag
      description: Sets the healthy flag of the application for demonstration purposes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HealthPayload'

      responses:
        '200':
          description: A booking based on search criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '500':
          description: Internal server error
  
  /health:
    get:
      summary: Health check
      description: Check the health status of the API.
      responses:
        '200':
          description: API is healthy
        '400':
          description: Incorrent parameters supplied
        '500':
          description: API is not healthy
          
components:
  schemas:
    HealthPayload:
      type: object
      properties:
        health:
          type: boolean
    Message:
      type: object
      properties:
        message:
          type: string
          example: "Success"
    BookingResult:
      type: object
      properties:
        bookingCode:
          type: string
        contactDetails:
          type: array
          items:
            type: object
            properties:
              "@class":
                type: string
              address:
                type: string
            required:
            - "@class"
            - address
        itinerary:
          type: object
          properties:
            type:
              type: string
            connections:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                  duration:
                    type: string
                  origin:
                    type: object
                    properties:
                      IATACode:
                        type: string
                      name:
                        type: string
                      city:
                        type: object
                        properties:
                          IATACode:
                            type: string
                          name:
                            type: string
                          country:
                            type: object
                            properties:
                              code:
                                type: string
                              name:
                                type: string
                            required:
                            - code
                            - name
                        required:
                        - IATACode
                        - name
                        - country
                    required:
                    - IATACode
                    - name
                    - city
                  destination:
                    type: object
                    properties:
                      IATACode:
                        type: string
                      name:
                        type: string
                      city:
                        type: object
                        properties:
                          IATACode:
                            type: string
                          name:
                            type: string
                          country:
                            type: object
                            properties:
                              code:
                                type: string
                              name:
                                type: string
                            required:
                            - code
                            - name
                        required:
                        - IATACode
                        - name
                        - country
                    required:
                    - IATACode
                    - name
                    - city
                  segments:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        type:
                          type: string
                        informational:
                          type: boolean
                        departFrom:
                          type: object
                          properties:
                            IATACode:
                              type: string
                            name:
                              type: string
                            city:
                              type: object
                              properties:
                                IATACode:
                                  type: string
                                name:
                                  type: string
                                country:
                                  type: object
                                  properties:
                                    code:
                                      type: string
                                    name:
                                      type: string
                                  required:
                                  - code
                                  - name
                              required:
                              - IATACode
                              - name
                              - country
                          required:
                          - IATACode
                          - name
                          - city
                        arriveOn:
                          type: object
                          properties:
                            IATACode:
                              type: string
                            name:
                              type: string
                            city:
                              type: object
                              properties:
                                IATACode:
                                  type: string
                                name:
                                  type: string
                                country:
                                  type: object
                                  properties:
                                    code:
                                      type: string
                                    name:
                                      type: string
                                  required:
                                  - code
                                  - name
                              required:
                              - IATACode
                              - name
                              - country
                          required:
                          - IATACode
                          - name
                          - city
                        marketingFlight:
                          type: object
                          properties:
                            number:
                              type: string
                            carrier:
                              type: object
                              properties:
                                code:
                                  type: string
                                name:
                                  type: string
                              required:
                              - code
                              - name
                            status:
                              type: object
                              properties:
                                code:
                                  type: string
                                name:
                                  type: string
                              required:
                              - code
                              - name
                            numberOfStops:
                              type: integer
                            sellingClass:
                              type: object
                              properties:
                                code:
                                  type: string
                              required:
                              - code
                            operatingFlight:
                              type: object
                              properties:
                                number:
                                  type: string
                                carrier:
                                  type: object
                                  properties:
                                    code:
                                      type: string
                                    name:
                                      type: string
                                  required:
                                  - code
                                  - name
                                duration:
                                  type: string
                                flown:
                                  type: boolean
                                checkInStart:
                                  type: string
                                localCheckInStart:
                                  type: string
                                checkInEnd:
                                  type: string
                                localCheckInEnd:
                                  type: string
                                scheduledArrival:
                                  type: string
                                localScheduledArrival:
                                  type: string
                                scheduledDeparture:
                                  type: string
                                localScheduledDeparture:
                                  type: string
                                arrivalTerminal:
                                  type: object
                                  properties:
                                    name:
                                      type: string
                                  required:
                                  - name
                                cabin:
                                  type: object
                                  properties:
                                    code:
                                      type: string
                                    name:
                                      type: string
                                  required:
                                  - code
                                  - name
                                equipment:
                                  type: object
                                  properties:
                                    code:
                                      type: string
                                    name:
                                      type: string
                                  required:
                                  - code
                                  - name
                              required:
                              - number
                              - carrier
                              - duration
                              - flown
                              - checkInStart
                              - localCheckInStart
                              - checkInEnd
                              - localCheckInEnd
                              - scheduledArrival
                              - localScheduledArrival
                              - scheduledDeparture
                              - localScheduledDeparture
                              - arrivalTerminal
                              - cabin
                              - equipment
                          required:
                          - number
                          - carrier
                          - status
                          - numberOfStops
                          - sellingClass
                          - operatingFlight
                      required:
                      - id
                      - type
                      - informational
                      - departFrom
                      - arriveOn
                      - marketingFlight
                required:
                - id
                - duration
                - origin
                - destination
                - segments
          required:
          - type
          - connections
        passengers:
          type: object
          properties:
            id:
              type: integer
            firstName:
              type: string
            lastName:
              type: string
            title:
              type: object
              properties:
                code:
                  type: string
                name:
                  type: string
              required:
              - code
              - name
          required:
          - id
          - firstName
          - lastName
          - title
      required:
      - bookingCode
      - contactDetails
      - itinerary
      - passengers